+++ examples/custom_strategy.rs
@@ -1,6 +1,6 @@
 use ninterp::data::InterpData2D;
 use ninterp::prelude::*;
-use ninterp::strategy::*;
+use ninterp::strategy::traits::*;
 
 // Note: ninterp also re-exposes the internally used `ndarray` crate
 // `use ninterp::ndarray;`
