+++ src/strategy/mod.rs
@@ -2,180 +2,8 @@
 
 use super::*;
 
-mod cubic;
-pub use cubic::*;
-
-pub trait Strategy1D<D>: Debug + DynClone
-where
-    D: Data + RawDataClone + Clone,
-    D::Elem: PartialEq + Debug,
-{
-    fn init(&mut self, _data: &InterpData1D<D>) -> Result<(), ValidateError> {
-        Ok(())
-    }
-
-    fn interpolate(
-        &self,
-        data: &InterpData1D<D>,
-        point: &[D::Elem; 1],
-    ) -> Result<D::Elem, InterpolateError>;
-
-    /// Does this type's [`Strategy1D::interpolate`] provision for extrapolation?
-    fn allow_extrapolate(&self) -> bool;
-}
-
-clone_trait_object!(<D> Strategy1D<D>);
-
-impl<D> Strategy1D<D> for Box<dyn Strategy1D<D>>
-where
-    D: Data + RawDataClone + Clone,
-    D::Elem: PartialEq + Debug,
-{
-    fn init(&mut self, data: &InterpData1D<D>) -> Result<(), ValidateError> {
-        (**self).init(data)
-    }
-
-    fn interpolate(
-        &self,
-        data: &InterpData1D<D>,
-        point: &[D::Elem; 1],
-    ) -> Result<D::Elem, InterpolateError> {
-        (**self).interpolate(data, point)
-    }
-
-    fn allow_extrapolate(&self) -> bool {
-        (**self).allow_extrapolate()
-    }
-}
-
-pub trait Strategy2D<D>: Debug + DynClone
-where
-    D: Data + RawDataClone + Clone,
-    D::Elem: PartialEq + Debug,
-{
-    fn init(&mut self, _data: &InterpData2D<D>) -> Result<(), ValidateError> {
-        Ok(())
-    }
-
-    fn interpolate(
-        &self,
-        data: &InterpData2D<D>,
-        point: &[D::Elem; 2],
-    ) -> Result<D::Elem, InterpolateError>;
-
-    /// Does this type's [`Strategy2D::interpolate`] provision for extrapolation?
-    fn allow_extrapolate(&self) -> bool;
-}
-
-clone_trait_object!(<D> Strategy2D<D>);
-
-impl<D> Strategy2D<D> for Box<dyn Strategy2D<D>>
-where
-    D: Data + RawDataClone + Clone,
-    D::Elem: PartialEq + Debug,
-{
-    fn init(&mut self, data: &InterpData2D<D>) -> Result<(), ValidateError> {
-        (**self).init(data)
-    }
-
-    fn interpolate(
-        &self,
-        data: &InterpData2D<D>,
-        point: &[D::Elem; 2],
-    ) -> Result<D::Elem, InterpolateError> {
-        (**self).interpolate(data, point)
-    }
-
-    fn allow_extrapolate(&self) -> bool {
-        (**self).allow_extrapolate()
-    }
-}
-
-pub trait Strategy3D<D>: Debug + DynClone
-where
-    D: Data + RawDataClone + Clone,
-    D::Elem: PartialEq + Debug,
-{
-    fn init(&mut self, _data: &InterpData3D<D>) -> Result<(), ValidateError> {
-        Ok(())
-    }
-
-    fn interpolate(
-        &self,
-        data: &InterpData3D<D>,
-        point: &[D::Elem; 3],
-    ) -> Result<D::Elem, InterpolateError>;
-
-    /// Does this type's [`Strategy3D::interpolate`] provision for extrapolation?
-    fn allow_extrapolate(&self) -> bool;
-}
-
-clone_trait_object!(<D> Strategy3D<D>);
-
-impl<D> Strategy3D<D> for Box<dyn Strategy3D<D>>
-where
-    D: Data + RawDataClone + Clone,
-    D::Elem: PartialEq + Debug,
-{
-    fn init(&mut self, data: &InterpData3D<D>) -> Result<(), ValidateError> {
-        (**self).init(data)
-    }
-
-    fn interpolate(
-        &self,
-        data: &InterpData3D<D>,
-        point: &[D::Elem; 3],
-    ) -> Result<D::Elem, InterpolateError> {
-        (**self).interpolate(data, point)
-    }
-
-    fn allow_extrapolate(&self) -> bool {
-        (**self).allow_extrapolate()
-    }
-}
-
-pub trait StrategyND<D>: Debug + DynClone
-where
-    D: Data + RawDataClone + Clone,
-    D::Elem: PartialEq + Debug,
-{
-    fn init(&mut self, _data: &InterpDataND<D>) -> Result<(), ValidateError> {
-        Ok(())
-    }
-
-    fn interpolate(
-        &self,
-        data: &InterpDataND<D>,
-        point: &[D::Elem],
-    ) -> Result<D::Elem, InterpolateError>;
-
-    /// Does this type's [`StrategyND::interpolate`] provision for extrapolation?
-    fn allow_extrapolate(&self) -> bool;
-}
-
-clone_trait_object!(<D> StrategyND<D>);
-
-impl<D> StrategyND<D> for Box<dyn StrategyND<D>>
-where
-    D: Data + RawDataClone + Clone,
-    D::Elem: PartialEq + Debug,
-{
-    fn init(&mut self, data: &InterpDataND<D>) -> Result<(), ValidateError> {
-        (**self).init(data)
-    }
-
-    fn interpolate(
-        &self,
-        data: &InterpDataND<D>,
-        point: &[D::Elem],
-    ) -> Result<D::Elem, InterpolateError> {
-        (**self).interpolate(data, point)
-    }
-
-    fn allow_extrapolate(&self) -> bool {
-        (**self).allow_extrapolate()
-    }
-}
+pub mod cubic;
+pub mod traits;
 
 // This method contains code from RouteE Compass, another open-source NREL-developed tool
 // <https://www.nrel.gov/transportation/route-energy-prediction-model.html>
